split geom
